Modals can now have navBars and can also call push to navigate within the modal.
