+++ lib/reducer.js
@@ -4,7 +4,7 @@ const {
   StateUtils: NavigationStateUtils,
 } = NavigationExperimental;
 
-let createReducer = (currentState = {}, action = {}) => {
+const createReducer = (currentState = {}, action = {}) => {
   // Make sure that we don't have two modals on top of each other.
   if (action.type === 'modal') {
     if (action.key === currentState.activeKey) {
@@ -31,10 +31,22 @@ let createReducer = (currentState = {}, action = {}) => {
       return setActiveKey(NavigationStateUtils.jumpToIndex(state, action.index), action.key);
     }
     case 'push': {
-      let selectedTabState = currentState.tabState[currentState.selectedTab];
       let scene = findScene(action.key);
       let props = Object.assign({}, scene, { ...action.props });
       let key = action.key;
+
+      if (currentState.modal) {
+        let exists = currentState.modalState.children.filter(c => c.key === key || c.key.includes(`${key}_`));
+        if (exists.length) {
+          key = `${key}_${exists.length}`;
+        }
+        let newModalState = NavigationStateUtils.push(currentState.modalState, { key, props });
+        let navState = Object.assign({}, currentState, { modalState: newModalState });
+console.log(navState);
+        return setActiveKey(Object.assign({}, navState), action.key);
+      }
+
+      let selectedTabState = currentState.tabState[currentState.selectedTab];
       let exists = selectedTabState.children.filter(c => c.key === key || c.key.includes(`${key}_`));
       if (exists.length) {
         key = `${key}_${exists.length}`;
@@ -48,7 +60,8 @@ let createReducer = (currentState = {}, action = {}) => {
     case 'modal': {
       let scene = findScene(action.key);
       let props = Object.assign({}, scene, { ...action.props });
-      let navState = NavigationStateUtils.push(currentState, { key: action.key, props });
+      let modalState = NavigationStateUtils.push(currentState.modalState, { key: action.key, props });
+      let navState = Object.assign({}, currentState, { modalState });
       return setActiveKey(Object.assign({}, navState, { modal: true }), action.key);
     }
     case 'back':
@@ -63,10 +76,12 @@ let createReducer = (currentState = {}, action = {}) => {
         return setActiveKey(Object.assign({}, currentState, { tabState: newTabState }), activeKey);
       }
 
-      let navState = currentState.index > 0 ? NavigationStateUtils.pop(currentState) : currentState;
+      let modalState = NavigationStateUtils.pop(currentState.modalState);
+      let navState = Object.assign({}, currentState, { modalState });
       let selectedTab = navState.tabState[navState.selectedTab];
       let activeKey = selectedTab.children[selectedTab.index].key;
-      return setActiveKey(Object.assign({}, navState, { modal: false }), activeKey);
+
+      return setActiveKey(Object.assign({}, navState, { modal: navState.modalState.children.length > 0 }), activeKey);
     }
     case 'popTo': {
       if (!currentState.modal) {
