+++ lib/renderer.js
@@ -1,93 +1,80 @@
 import React, { Component, PropTypes } from 'react';
-import { Animated, Dimensions } from 'react-native';
+import { Animated, NavigationExperimental } from 'react-native';
 
-let { width, height } = Dimensions.get('window');
+import navBar from './navBar';
+import navigationHelper from './navigationHelper';
 
+const {
+  AnimatedView: NavigationAnimatedView,
+  Card: NavigationCard,
+  Container: NavigationContainer,
+} = NavigationExperimental;
 
 class Renderer extends Component {
   static propTypes = {
-    children: PropTypes.oneOfType([
-      PropTypes.array,
-      PropTypes.object,
-    ]),
-    direction: PropTypes.string,
-    show: PropTypes.bool,
+    navState: PropTypes.object,
+    dispatch: PropTypes.func,
   };
 
-  static defaultProps = {
-    direction: 'vertical',
-  };
-
-  constructor(props) {
-    super(props);
+  constructor(props, context) {
+    super(props, context);
 
     this.state = {
-      animatedValue: new Animated.Value(height),
-      children: props.children,
-      direction: props.direction,
-      show: props.show,
+      navState: props.navState,
     };
   }
-  componentWillMount() {
-    let animatedValue = this.state.direction === 'vertical' ? height : width;
-    let _animatedValue = this.props.show ? 0 : animatedValue;
-
-    Animated.spring(
-      this.state.animatedValue,
-      {
-        toValue: _animatedValue,
-        duration: 400,
-      },
-    ).start();
-  }
-  componentWillReceiveProps(nextProps) {
-    this.setState({
-      children: nextProps.children || this.state.children,
-      direction: nextProps.children === undefined ? this.state.direction : nextProps.direction,
-      show: nextProps.show,
-    });
-  }
-  shouldComponentUpdate(nextProps) {
-    if (nextProps.show !== this.state.show) {
-      return true;
+  componentWillReceiveProps(props) {
+    if (props.navState.index === this.state.navState.index) {
+      return;
     }
 
-    return false;
-  }
-  componentDidUpdate() {
-    let animatedValue = this.state.direction === 'vertical' ? height : width;
-    let _animatedValue = this.props.show ? 0 : animatedValue;
-
-    Animated.timing(
-      this.state.animatedValue,
-      {
-        toValue: _animatedValue,
-        duration: 400,
-      },
-    ).start(() => {
-      if (!this.state.show && this.state.children) {
-        this.setState({
-          children: null,
-        });
-        this.forceUpdate();
-      }
+    this.setState({
+      navState: props.navState,
     });
+  }
+  shouldComponentUpdate(nextProps, nextState) {
+    return nextState.navState.index !== this.state.navState.index;
   }
   render() {
-    let top = 0;
-    let left = 0;
-    if (this.state.direction === 'vertical') {
-      top = this.state.animatedValue;
-    } else {
-      left = this.state.animatedValue;
-    }
+    let duration = 300;
+
+    let animation = (pos, navState) => {
+      Animated.timing(pos, { toValue: navState.index, duration }).start();
+    };
 
     return (
-      <Animated.View style={[{ position: 'absolute', top, left, overflow: 'hidden', height, width }]}>
-        {this.state.children}
-      </Animated.View>
+      <NavigationAnimatedView
+        navigationState={this.state.navState}
+        onNavigate={this._handleNavigation}
+        applyAnimation={animation}
+        renderOverlay={navBar(navigationHelper(this._handleNavigation), this.props.dispatch)}
+        renderScene={this._renderCard}
+        style={{ flex: 1 }}
+      />
     );
   }
+  _renderCard = (props) => (
+    <NavigationCard
+      panHandlers={null}
+      {...props}
+      key={`card_${props.scene.navigationState.key}`}
+      renderScene={this._renderScene}
+    />
+  );
+  /* eslint no-use-before-define: 0 */
+  _renderScene = (props) => {
+    let { component, ...otherProps } = props.scene.navigationState.props;
+    let SceneComponent = component;
+
+    return (
+      <SceneComponent {...otherProps} navigate={navigationHelper(this._handleNavigation)} />
+    );
+  };
+  _handleNavigation = (action) => {
+    let { key, type, ...otherProps } = action;
+    let sceneAction = { key, type, props: otherProps };
+    return sceneAction;
+  };
 }
 
-export default Renderer;
+export default NavigationContainer.create(Renderer);
