+++ lib/tabController.js
@@ -2,8 +2,8 @@ import React, { Component, PropTypes } from 'react';
 import { Text, View } from 'react-native';
 import Tabs from './tabBar/tabBar';
 
-import TabView from './tabView';
 import Renderer from './renderer';
+import Modal from './modal';
 import navigationHelper from './navigationHelper';
 
 /* eslint no-use-before-define: 0 */
@@ -54,27 +54,35 @@ class TabController extends Component {
       }
     }
 
-    props.dispatch({ type: 'NAVINIT', state: {
-      key: 'tabController',
-      activeKey: initialScene,
-      index,
-      selectedTab,
-      children,
-      modal: initialSceneIsModal,
-      scenes: props.scenes,
-      tabState: this._tabs.map(t => {
-        let { key, ...otherProps } = t;
-
-        return {
+    props.dispatch({
+      type: 'NAVINIT',
+      state: {
+        activeKey: initialScene,
+        children,
+        index,
+        key: 'tabController',
+        modal: initialSceneIsModal,
+        modalState: {
+          children: [],
           index: 0,
-          children: [{
+          key: 'modal',
+        },
+        scenes: props.scenes,
+        selectedTab,
+        tabState: this._tabs.map(t => {
+          let { key, ...otherProps } = t;
+
+          return {
+            index: 0,
+            children: [{
+              key,
+              props: otherProps.props,
+            }],
             key,
-            props: otherProps.props,
-          }],
-          key,
-        };
-      }),
-    } });
+          };
+        }),
+      },
+    });
   }
   render() {
     if (!this.props.navState.children) {
@@ -89,12 +97,12 @@ class TabController extends Component {
 
     let currentChild = navigationState.children[navigationState.index];
     let scene;
-    if (currentChild.props.type === 'SCENE') {
-      let { component, ...otherProps } = currentChild.props;
-      let SceneComponent = component;
+
+    if (navigationState.modal) {
+      currentChild = navigationState.modalState.children[0];
       scene = (
         <View style={{ flex: 1 }}>
-          <SceneComponent {...otherProps} navigate={navigationHelper()} />
+          <Renderer navState={navigationState.modalState} dispatch={this.props.dispatch} />
         </View>
       );
     }
@@ -109,21 +117,19 @@ class TabController extends Component {
 
           return (
             <View key={index} style={style}>
-              <TabView navState={navigationState.tabState[index]} dispatch={this.props.dispatch} />
+              <Renderer navState={navigationState.tabState[index]} dispatch={this.props.dispatch} />
             </View>
           );
         })}
-        {currentChild.type !== 'SCENE' &&
-          <Tabs style={{ backgroundColor: 'white' }} onSelectTab={_tabSelect}>
-            {this._tabs.map((tab, index) => {
-              let tabRenderer = tab.props.icon || this._renderDefaultTab;
-              return tabRenderer(tab.props, index, tab.key, navigationState.selectedTab);
-            })}
-          </Tabs>
-        }
-        <Renderer show={currentChild.props.type === 'SCENE'} direction={currentChild.props.direction} scene={currentChild}>
+        <Tabs style={{ backgroundColor: 'white' }} onSelectTab={_tabSelect}>
+          {this._tabs.map((tab, index) => {
+            let tabRenderer = tab.props.icon || this._renderDefaultTab;
+            return tabRenderer(tab.props, index, tab.key, navigationState.selectedTab);
+          })}
+        </Tabs>
+        <Modal show={navigationState.modal} navState={navigationState.modalState} direction={currentChild.props.direction} scene={currentChild}>
           {scene}
-        </Renderer>
+        </Modal>
       </View>
     );
   };
