+++ package.json
@@ -2,7 +2,7 @@
   "name": "react-native-router-ex",
   "author": "Jarred Witt",
   "description": "React Native router that uses NavigationExperimental.",
-  "version": "2.7.0",
+  "version": "2.8.0",
   "repository": {
     "type": "git",
     "url": "https://github.com/jarredwitt/react-native-router-ex"
@@ -36,12 +36,12 @@
     "chai-as-promised": "^5.2.0",
     "enzyme": "^2.1.0",
     "eslint": "^2.9.0",
-    "eslint-config-airbnb": "^8.0.0",
-    "eslint-plugin-import": "^1.6.1",
-    "eslint-plugin-jsx-a11y": "^1.0.4",
+    "eslint-config-airbnb": "^9.0.1",
+    "eslint-plugin-import": "^1.7.0",
+    "eslint-plugin-jsx-a11y": "^1.2.0",
     "eslint-plugin-react": "^5.0.1",
     "estraverse-fb": "^1.3.1",
-    "expect": "^1.19.0",
+    "expect": "^1.20.1",
     "mocha": "^2.4.5",
     "precommit-hook-eslint": "^3.0.0",
     "react": "^15.0.2",
